rand utile
...
